+++ xclip/xclip.h
@@ -4,6 +4,8 @@
 #define XCLIP_IN 0
 #define XCLIP_OUT 1
 
+#include <stdint.h>
+
 /*
  * Get or set clipboard data
  * Allocates a memory buffer and set the pointer to buf_ptr in get mode.
@@ -12,6 +14,6 @@
  * Returns 0 on success.
  * Returns -1 if an error occured.
  */
-extern int xclip_util(int io, const char *atom_name, unsigned long *len_ptr, char **buf_ptr);
+extern int xclip_util(int io, const char *atom_name, uint32_t *len_ptr, char **buf_ptr);
 
 #endif  // XCLIP_XCLIP_H_
